requirements.txt
fastapi
motor
pydantic
uvicorn
python-dotenv
pytest